refactor: add recipes folder
